+++ src/azure-cli/azure/cli/command_modules/vm/custom.py
@@ -249,8 +249,13 @@ def create_managed_disk(cmd, resource_group_name, disk_name, location=None,  # p
                         # below are generated internally from 'source'
                         source_blob_uri=None, source_disk=None, source_snapshot=None,
                         source_storage_account_id=None, no_wait=False, tags=None, zone=None,
-                        disk_iops_read_write=None, disk_mbps_read_write=None, hyper_v_generation=None):
-    Disk, CreationData, DiskCreateOption = cmd.get_models('Disk', 'CreationData', 'DiskCreateOption')
+                        disk_iops_read_write=None, disk_mbps_read_write=None, hyper_v_generation=None,
+                        encryption_type=None, disk_encryption_set=None):
+    from msrestazure.tools import resource_id, is_valid_resource_id
+    from azure.cli.core.commands.client_factory import get_subscription_id
+
+    Disk, CreationData, DiskCreateOption, Encryption = cmd.get_models(
+        'Disk', 'CreationData', 'DiskCreateOption', 'Encryption')
 
     location = location or _get_resource_group_location(cmd.cli_ctx, resource_group_name)
     if source_blob_uri:
@@ -263,8 +268,6 @@ def create_managed_disk(cmd, resource_group_name, disk_name, location=None,  # p
         option = DiskCreateOption.empty
 
     if source_storage_account_id is None and source_blob_uri is not None:
-        from azure.cli.core.commands.client_factory import get_subscription_id
-        from msrestazure.tools import resource_id
         subscription_id = get_subscription_id(cmd.cli_ctx)
         storage_account_name = source_blob_uri.split('.')[0].split('/')[-1]
         source_storage_account_id = resource_id(
@@ -282,8 +285,21 @@ def create_managed_disk(cmd, resource_group_name, disk_name, location=None,  # p
 
     if size_gb is None and upload_size_bytes is None and (option == DiskCreateOption.empty or for_upload):
         raise CLIError('usage error: --size-gb or --upload-size-bytes required to create an empty disk')
+
+    if disk_encryption_set is not None and not is_valid_resource_id(disk_encryption_set):
+        disk_encryption_set = resource_id(
+            subscription=get_subscription_id(cmd.cli_ctx), resource_group=resource_group_name,
+            namespace='Microsoft.Compute', type='diskEncryptionSets', name=disk_encryption_set)
+
+    if disk_encryption_set is not None and encryption_type is None:
+        raise CLIError('usage error: Please specify --encryption-type.')
+    if encryption_type is not None:
+        encryption = Encryption(type=encryption_type, disk_encryption_set_id=disk_encryption_set)
+    else:
+        encryption = None
+
     disk = Disk(location=location, creation_data=creation_data, tags=(tags or {}),
-                sku=_get_sku_object(cmd, sku), disk_size_gb=size_gb, os_type=os_type)
+                sku=_get_sku_object(cmd, sku), disk_size_gb=size_gb, os_type=os_type, encryption=encryption)
 
     if hyper_v_generation:
         disk.hyper_vgeneration = hyper_v_generation
@@ -532,7 +548,8 @@ def create_vm(cmd, vm_name, resource_group_name, image=None, size='Standard_DS1_
               identity_role='Contributor', identity_role_id=None, application_security_groups=None, zone=None,
               boot_diagnostics_storage=None, ultra_ssd_enabled=None, ephemeral_os_disk=None,
               proximity_placement_group=None, dedicated_host=None, dedicated_host_group=None, aux_subscriptions=None,
-              priority=None, max_price=None, eviction_policy=None, enable_agent=None, workspace=None, vmss=None):
+              priority=None, max_price=None, eviction_policy=None, enable_agent=None, workspace=None, vmss=None,
+              os_disk_encryption_set=None, data_disk_encryption_sets=None):
     from azure.cli.core.commands.client_factory import get_subscription_id
     from azure.cli.core.util import random_string, hash_string
     from azure.cli.core.commands.arm import ArmTemplateBuilder
@@ -545,9 +562,23 @@ def create_vm(cmd, vm_name, resource_group_name, image=None, size='Standard_DS1_
                                                                 build_vm_daExtensionName_resource)
     from msrestazure.tools import resource_id, is_valid_resource_id
 
+    subscription_id = get_subscription_id(cmd.cli_ctx)
+
+    if os_disk_encryption_set is not None and not is_valid_resource_id(os_disk_encryption_set):
+        os_disk_encryption_set = resource_id(
+            subscription=subscription_id, resource_group=resource_group_name,
+            namespace='Microsoft.Compute', type='diskEncryptionSets', name=os_disk_encryption_set)
+
+    if data_disk_encryption_sets is None:
+        data_disk_encryption_sets = []
+    for i, des in enumerate(data_disk_encryption_sets):
+        if des is not None and not is_valid_resource_id(des):
+            data_disk_encryption_sets[i] = resource_id(
+                subscription=subscription_id, resource_group=resource_group_name,
+                namespace='Microsoft.Compute', type='diskEncryptionSets', name=des)
+
     storage_sku = disk_info['os'].get('storageAccountType')
 
-    subscription_id = get_subscription_id(cmd.cli_ctx)
     network_id_template = resource_id(
         subscription=subscription_id, resource_group=resource_group_name,
         namespace='Microsoft.Network')
@@ -660,7 +691,8 @@ def create_vm(cmd, vm_name, resource_group_name, image=None, size='Standard_DS1_
         boot_diagnostics_storage_uri=boot_diagnostics_storage, ultra_ssd_enabled=ultra_ssd_enabled,
         proximity_placement_group=proximity_placement_group, computer_name=computer_name,
         dedicated_host=dedicated_host, priority=priority, max_price=max_price, eviction_policy=eviction_policy,
-        enable_agent=enable_agent, vmss=vmss)
+        enable_agent=enable_agent, vmss=vmss, os_disk_encryption_set=os_disk_encryption_set,
+        data_disk_encryption_sets=data_disk_encryption_sets)
 
     vm_resource['dependsOn'] = vm_dependencies
 
@@ -2007,7 +2039,8 @@ def create_vmss(cmd, vmss_name, resource_group_name, image=None,
                 identity_role_id=None, zones=None, priority=None, eviction_policy=None,
                 application_security_groups=None, ultra_ssd_enabled=None, ephemeral_os_disk=None,
                 proximity_placement_group=None, aux_subscriptions=None, terminate_notification_time=None,
-                max_price=None, computer_name_prefix=None, orchestration_mode='ScaleSetVM', scale_in_policy=None):
+                max_price=None, computer_name_prefix=None, orchestration_mode='ScaleSetVM', scale_in_policy=None,
+                os_disk_encryption_set=None, data_disk_encryption_sets=None):
     from azure.cli.core.commands.client_factory import get_subscription_id
     from azure.cli.core.util import random_string, hash_string
     from azure.cli.core.commands.arm import ArmTemplateBuilder
@@ -2028,6 +2061,20 @@ def create_vmss(cmd, vmss_name, resource_group_name, image=None,
         storage_sku = disk_info['os'].get('storageAccountType')
 
         subscription_id = get_subscription_id(cmd.cli_ctx)
+
+        if os_disk_encryption_set is not None and not is_valid_resource_id(os_disk_encryption_set):
+            os_disk_encryption_set = resource_id(
+                subscription=subscription_id, resource_group=resource_group_name,
+                namespace='Microsoft.Compute', type='diskEncryptionSets', name=os_disk_encryption_set)
+
+        if data_disk_encryption_sets is None:
+            data_disk_encryption_sets = []
+        for i, des in enumerate(data_disk_encryption_sets):
+            if des is not None and not is_valid_resource_id(des):
+                data_disk_encryption_sets[i] = resource_id(
+                    subscription=subscription_id, resource_group=resource_group_name,
+                    namespace='Microsoft.Compute', type='diskEncryptionSets', name=des)
+
         network_id_template = resource_id(
             subscription=subscription_id, resource_group=resource_group_name,
             namespace='Microsoft.Network')
@@ -2220,7 +2267,8 @@ def create_vmss(cmd, vmss_name, resource_group_name, image=None,
             eviction_policy=eviction_policy, application_security_groups=application_security_groups,
             ultra_ssd_enabled=ultra_ssd_enabled, proximity_placement_group=proximity_placement_group,
             terminate_notification_time=terminate_notification_time, max_price=max_price,
-            scale_in_policy=scale_in_policy)
+            scale_in_policy=scale_in_policy, os_disk_encryption_set=os_disk_encryption_set,
+            data_disk_encryption_sets=data_disk_encryption_sets)
 
         vmss_resource['dependsOn'] = vmss_dependencies
 
@@ -2953,4 +3001,42 @@ def execute_query_for_vm(cmd, client, resource_group_name, vm_name, analytics_qu
         raise CLIError('Cannot find the corresponding log analytics workspace. '
                        'Please check the status of log analytics workpsace.')
     return client.query(workspace, QueryBody(query=analytics_query, timespan=timespan))
+# endregion
+
+
+# disk encryption set
+def create_disk_encryption_set(cmd, client, resource_group_name, disk_encryption_set_name,
+                               key_url, source_vault, location=None, tags=None, no_wait=False):
+    from msrestazure.tools import resource_id, is_valid_resource_id
+    DiskEncryptionSet, EncryptionSetIdentity, KeyVaultAndKeyReference, SourceVault = cmd.get_models(
+        'DiskEncryptionSet', 'EncryptionSetIdentity', 'KeyVaultAndKeyReference', 'SourceVault')
+    encryption_set_identity = EncryptionSetIdentity(type='SystemAssigned')
+    if not is_valid_resource_id(source_vault):
+        source_vault = resource_id(subscription=client.config.subscription_id, resource_group=resource_group_name,
+                                   namespace='Microsoft.KeyVault', type='vaults', name=source_vault)
+    source_vault = SourceVault(id=source_vault)
+    keyVault_and_key_reference = KeyVaultAndKeyReference(source_vault=source_vault, key_url=key_url)
+    disk_encryption_set = DiskEncryptionSet(location=location, tags=tags, identity=encryption_set_identity,
+                                            active_key=keyVault_and_key_reference)
+    return sdk_no_wait(no_wait, client.create_or_update, resource_group_name, disk_encryption_set_name,
+                       disk_encryption_set)
+
+
+def list_disk_encryption_sets(cmd, client, resource_group_name=None):
+    if resource_group_name:
+        return client.list_by_resource_group(resource_group_name)
+    return client.list()
+
+
+def update_disk_encryption_set(instance, client, resource_group_name, key_url=None, source_vault=None):
+    from msrestazure.tools import resource_id, is_valid_resource_id
+    if not is_valid_resource_id(source_vault):
+        source_vault = resource_id(subscription=client.config.subscription_id, resource_group=resource_group_name,
+                                   namespace='Microsoft.KeyVault', type='vaults', name=source_vault)
+    if key_url:
+        instance.active_key.key_url = key_url
+    if source_vault:
+        instance.active_key.source_vault.id = source_vault
+    return instance
+
 # endregion